+++ R/sparse-spectrum.R
@@ -1,7 +1,8 @@
 sparse_spectrum <- function(x, y,
                             x_unit, y_unit,
                             label = "sparse spectrum",
-                            x_lab = x_unit, y_lab = y_unit) {
+                            x_lab = x_unit, y_lab = y_unit,
+                            labels = NULL) {
   checkmate::qassert(x, "N")
   checkmate::qassert(y, "N")
   checkmate::qassert(x_unit, "S1")
@@ -17,6 +18,7 @@ sparse_spectrum <- function(x, y,
   attr(df, "x_lab") <- x_lab
   attr(df, "y_lab") <- y_lab
   class(df) <- c("sparse_spectrum", "data.frame")
+  if (!is.null(labels)) df <- set_labels(df, labels)
   df
 }
 
@@ -29,37 +31,78 @@ print.sparse_spectrum <- function(x, ...) {
       label(x), "\n",
       "  size = ", nrow(x), "\n",
       "  x = ", x_unit(x), "\n",
-      "  y = ", y_unit(x), "\n\n",
+      "  y = ", y_unit(x), "\n",
+      if (!is.null(x$labels)) "  (has labels)\n",
+      "\n",
       sep = ""
     )
 }
 
 #' @export
 as.data.frame.sparse_spectrum <- function(x, ...) {
-  data.frame(x = x$x,
-             y = x$y)
+  df <- data.frame(x = x$x, y = x$y)
+  if (!is.null(x$labels)) df$labels <- x$labels
+  df
+}
+
+#' Add labels
+#'
+#' Adds labels to an object, typically for plotting purposes.
+#'
+#' @param x Object.
+#'
+#' @param labels Character vector of labels to add to the object.
+#' In the case of sparse spectra, there should be one label for each partial.
+#'
+#' @return The original object, with labels added.
+#'
+#' @rdname set_labels
+#'
+#' @examples
+#' spectrum <- sparse_pi_spectrum("60 64 67", num_harmonics = 1)
+#' labels <- as.character(c(1, 2, 3))
+#' spectrum_with_labels <- set_labels(spectrum, labels)
+#' plot(spectrum_with_labels)
+#'
+#' @export
+set_labels <- function(x, labels) {
+  UseMethod("set_labels")
+}
+
+#' @rdname set_labels
+#' @export
+set_labels.sparse_spectrum <- function(x, labels) {
+  checkmate::qassert(labels, sprintf("S%i", nrow(x)))
+  x$labels <- labels
+  x
 }
 
 #' @export
 plot.sparse_spectrum <- function(x, ggplot = FALSE, xlim = NULL, ...) {
   df <- as.data.frame(x)
-  n <- nrow(df)
-  df2 <- data.frame(x = numeric(n * 3), y = numeric(n * 3))
-  for (i in seq_len(n)) {
-    I <- (i - 1L) * 3L
-    df2$x[I + 1:3] <- df$x[i]
-    df2$y[I + 2L] <- df$y[i]
-  }
   if (ggplot) {
     assert_installed("ggplot2")
-    tibble::tibble(x = df2$x, y = df2$y) %>%
-      ggplot2::ggplot(ggplot2::aes_string(x = "x", y = "y")) +
-      ggplot2::geom_line() +
+      ggplot2::ggplot(df, ggplot2::aes_string(x = "x", xend = "x",
+                                              y = 0, yend = "y")) +
+      ggplot2::geom_segment() +
       ggplot2::scale_x_continuous(x_lab(x), limits = xlim) +
       ggplot2::scale_y_continuous(y_lab(x))
   } else {
+    n <- nrow(df)
+    df2 <- data.frame(x = numeric(n * 3), y = numeric(n * 3))
+    for (i in seq_len(n)) {
+      I <- (i - 1L) * 3L
+      df2$x[I + 1:3] <- df$x[i]
+      df2$y[I + 2L] <- df$y[i]
+    }
     plot(df2$x, df2$y, xlab = x_lab(x), ylab = y_lab(x),
          type = "l", xlim = xlim, ...)
+    if (!is.null(df$labels)) {
+      for (i in seq_len(nrow(df))) {
+        graphics::text(df$x[i], 0, rep("\u2588", times = nchar(df$label[i])) %>% paste(collapse = ""), col = "white")
+        graphics::text(df$x[i], 0, df$label[i])
+      }
+    }
   }
 }
 
@@ -159,7 +202,7 @@ combine_sparse_spectra <- function(..., digits = 6) {
     collapse_summing_amplitudes(digits = digits) %>%
     {
       f <- if (octave_invariant) .sparse_pc_spectrum else .sparse_pi_spectrum
-      f(.$x, .$y)
+      f(.$x, .$y, labels = .$labels)
     }
 
   if (output_class == "sparse_fr_spectrum") sparse_fr_spectrum(res) else res
@@ -169,8 +212,10 @@ collapse_summing_amplitudes <- function(x, digits, modulo = NA_real_) {
   checkmate::qassert(modulo, "n1(0,)")
   if (!is.list(x) ||
       !all(purrr::map_lgl(x, ~ is.data.frame(.) &&
-                          identical(names(.), c("x", "y")))))
+                          all(c("x", "y") %in% names(.)))))
     stop("x must be a list of data frames with columns 'x' and 'y'")
+  has_labels <- !is.null(x[[1]]$labels)
+
   x %>%
     data.table::rbindlist() %>%
     {
@@ -180,9 +225,23 @@ collapse_summing_amplitudes <- function(x, digits, modulo = NA_real_) {
     } %>%
     {reduce_by_key(
       keys = .$x,
-      values = .$y,
-      function(x, y) sum_amplitudes(x, y, coherent = FALSE),
+      values = if (has_labels) purrr::map2(.$y, .$labels, ~ list(amplitude = .x, label = .y)) else .$y,
+      function(x, y) {
+        if (has_labels) {
+          list(amplitude = sum_amplitudes(x, y$amplitude, coherent = FALSE),
+               label = y$label)
+        } else {
+          sum_amplitudes(x, y, coherent = FALSE)
+        }
+      },
       key_type = "numeric"
-    )} %>%
-    magrittr::set_names(c("x", "y"))
+    )} %>% {
+      if (has_labels) {
+        list(x = .[[1]],
+             y = purrr::map_dbl(.[[2]], "amplitude"),
+             labels = purrr::map_chr(.[[2]], "label"))
+      } else {
+        magrittr::set_names(., c("x", "y"))
+      }
+    }
 }